@@ -2,39 +2,42 @@
   "name": "app",
   "version": "0.0.0",
   "license": "MIT",
-  "bin": "cli.js",
   "author": "Martin",
+  "type":"module",
   "scripts": {
     "test": "xo && ava",
-    "start": "node cli.js"
+    "start": "ts-node --esm cli.ts"
   },
   "dependencies": {
-    "@babel/cli": "^7.6.4",
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "@babel/preset-react": "^7.6.3",
-    "@babel/register": "^7.6.2",
+    "@babel/cli": "^7.22.15",
+    "@babel/core": "^7.22.20",
+    "@babel/preset-env": "^7.22.20",
+    "@babel/preset-react": "^7.22.15",
+    "@babel/register": "^7.22.15",
+    "@types/react": "^18.2.21",
     "ava": "^2.4.0",
     "babel-plugin-transform-export-extensions": "^6.22.0",
     "chalk": "^2.4.2",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-xo-react": "^0.20.0",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^2.2.0",
-    "import-jsx": "^3.0.0",
-    "ink": "^2.5.0",
-    "ink-big-text": "^1.0.1",
-    "ink-box": "^1.0.0",
-    "ink-gradient": "^1.0.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^2.5.1",
+    "import-jsx": "^5.0.0",
+    "ink": "^4.4.1",
+    "ink-big-text": "^2.0.0",
+    "ink-box": "^2.0.0",
+    "ink-gradient": "^3.0.0",
     "ink-progress-bar": "^3.0.0",
-    "ink-spinner": "^3.0.1",
-    "ink-testing-library": "^1.0.2",
-    "meow": "^5.0.0",
-    "node-fetch": "^2.6.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.11.0",
-    "rxjs": "^6.5.3",
-    "xo": "^0.25.3"
+    "ink-spinner": "^5.0.0",
+    "ink-testing-library": "^3.0.0",
+    "meow": "^12.1.1",
+    "node-fetch": "^2.7.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "rxjs": "^6.6.7",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2",
+    "xo": "^0.25.4"
   },
   "ava": {
     "require": [
